requirements.txt
psycopg2==2.7
pytz==2016.10
redis==2.10.5
celery==4.0.2
Flask==0.12
Flask-Cors==3.0.2
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.2
PyJWT==1.5.0
cryptography==1.8.1
passlib==1.7.1
cffi
influxdb
sympy==1.1
docopt
gunicorn